issue #45: Fix traditional background colour sequence
There is an error in the colouring for the ANSI/VT100 sequences for
ESC [40m to [47m that incorrectly sets foreground colour. These should
set the background colour.

The same issue applies for ESC [100m to ESC [107m for bright background.

This patch fixes the support for the traditional 8 colour (plus 8
bright colour) support.